# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A PLATFORM OPERATIONS API in golang using Gin framework",
        "title": "PLATFORM OPERATIONS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/request": {
            "post": {
                "description": "Create a Request and saves it in postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new Request",
                "parameters": [
                    {
                        "description": "Create Request",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateRequestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerRequest"
                        }
                    }
                }
            }
        },
        "/request/{id}": {
            "delete": {
                "description": "Get a Request from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "delete a Request with a specified id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update a Request from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "update a Request with a specified id",
                "parameters": [
                    {
                        "description": "update Request",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateRequestRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/requests": {
            "get": {
                "description": "List Requests according to the page_id and page_size",
                "produces": [
                    "application/json"
                ],
                "summary": "List all Requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/template": {
            "post": {
                "description": "Create a Template and saves it in postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new Template",
                "parameters": [
                    {
                        "description": "Create Template",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerTemplate"
                        }
                    }
                }
            }
        },
        "/template/{id}": {
            "get": {
                "description": "Get a Template according to their ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Get a Template from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "delete a Template with a specified id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update a Template from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "update a Template with a specified id",
                "parameters": [
                    {
                        "description": "update Template",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateTemplateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/templates": {
            "get": {
                "description": "List Templates according to the page_id and page_size",
                "produces": [
                    "application/json"
                ],
                "summary": "List all Templates",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "api.CreateRequestRequest": {
            "type": "object",
            "properties": {
                "layer_id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.CreateTemplateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "git_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.SwaggerRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "layer_id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "api.SwaggerTemplate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdateRequestRequest": {
            "type": "object",
            "properties": {
                "layer_id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.UpdateTemplateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "git_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}